consumer/requirement.txt
newspaper3k
kafka-python
nltk
pymongo
